[ProgressTracker#743] Added dynamic sizing.
