+++ core/Sources/Components/ProgressTracker/View/SwiftUI/ProgressTrackerTrackView.swift
@@ -8,19 +8,6 @@
 
 import SwiftUI
 
-struct ProgressTrackerSizePreferences: PreferenceKey {
-    typealias Value = [Int: CGRect]
-
-    static var defaultValue = [Int: CGRect]()
-
-    static func reduce(
-        value: inout [Int: CGRect],
-        nextValue: () -> [Int: CGRect]
-    ) {
-        value.merge(nextValue()) { $1 }
-    }
-}
-
 struct ProgressTrackerTrackView: View {
     @ObservedObject private var viewModel: ProgressTrackerTrackViewModel
     private let orientation: ProgressTrackerOrientation
